fix: update picture name duplicate error

+++ src/controllers/lens-control.controller.ts
@@ -213,6 +213,10 @@ export class LensControlController {
 
     } else if (lens.partNo != oldLen.partNo) {  // if not update pic but update the partNo,
       // change old pic name to new partNo
+
+      if (fs.existsSync('./public/lensPic/' + lens.partNo + '.png', () => { throw new HttpErrors.BadRequest() })) {
+        throw new HttpErrors.BadRequest('料號重復')
+      }
       try {
         fs.rename('./public' + oldLen.url, './public/lensPic/' + lens.partNo + '.png', () => { });
       } catch{
